React Router for creating new projects done; editing functions in progress
