+++ package.json
@@ -14,9 +14,12 @@
   },
   "dependencies": {
     "assistant-frontend": "file:",
+    "localforage": "^1.10.0",
+    "match-sorter": "^6.3.4",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
-    "react-router-dom": "^6.26.0"
+    "react-router-dom": "^6.26.0",
+    "sort-by": "^1.2.0"
   },
   "devDependencies": {
     "@types/react": "^18.3.3",
