+++ src/App.jsx
@@ -1,5 +1,5 @@
 import { useState } from 'react'
-import { Outlet, Link, useLoaderData } from 'react-router-dom'
+import { Outlet, Link } from 'react-router-dom'
 import './App.css'
 import SideNavBar from './components/navbar/NavBar'
 import HomePage from './components/home-page/HomePage'
@@ -10,7 +10,7 @@ function App() {
     <div className='h-screen flex'>
       <SideNavBar />
       <div className='w-full h-screen flex'>
-        <Outlet />
+      <Outlet />
       </div>
     </div>
   )
