+++ src/components/navbar/AddProjectButton.jsx
@@ -7,11 +7,9 @@ export default function AddProjectButton({onClick, type = 'button'}) {
 
 
   return (
-    <div>
         <button type={type} onClick={onClick} className='nav-button'>
             <div className='gg-add-r'></div>
         </button>
-    </div>
   );
 }
