+++ src/components/navbar/SelectProjectButton.jsx
@@ -7,8 +7,8 @@ export default function SelectProjectButton() {
 
 
   return (
-    <div>
-        <button className='nav-button'>
+    <div className="flex flex-col items-center">
+        <button className='nav-button flex items-center justify-center'>
             <div className='gg-ereader' ></div>
         </button>
         <p className='caption'>Product Backlog</p>
